requirements.txt
pandas
-e .
numpy 
fastapi
uvicorn
mysql-connector-python